requirements.txt
pymongo==4.8.0
flask==3.0.3
flask_cors==5.0.0
python-dotenv==1.0.1
